+++ inc/Newton.h
@@ -5,16 +5,16 @@
 struct NewtonInitializer;
 extern NewtonInitializer newtonInitializer;
 
-void NewtonResetBounds(double *lx,
-                       double *rx,
-                       double *ty,
-                       double *by);
+void NewtonResetBounds(long double *lx,
+                       long double *rx,
+                       long double *ty,
+                       long double *by);
 
 
 
-void NewtonFillLine(double lx,
-                    double dx,
-                    double y,
+void NewtonFillLine(long double lx,
+                    long double dx,
+                    long double y,
                     uint32_t yix,
                     uint32_t width,
                     uint32_t *dst);
