+++ src/Newton.cpp
@@ -6,8 +6,8 @@
 #define MAX_ITERS_COUNT 0xff
 
 static uint32_t m_colorTable[MAX_ITERS_COUNT+1];
-static uint32_t getColor(double cx,
-                         double cy);
+static uint32_t getColor(long double cx,
+                         long double cy);
 
 void
 initColorTable() {
@@ -23,15 +23,15 @@ initColorTable() {
 
 // f(x) = sin(x) - 1
 // f'(x) = cos(x)
-uint32_t getColor( double zx,
-                   double zy ) {
+uint32_t getColor( long double zx,
+                   long double zy ) {
   uint32_t iterCount = 0;
-  static const double maxT = 1e+6;
-  static const double minT = 1e-6;
+  static const long double maxT = 1e+6;
+  static const long double minT = 1e-6;
   
-  std::complex<double> z(zx, zy);
-  std::complex<double> t = z;
-  std::complex<double> d = t;
+  std::complex<long double> z(zx, zy);
+  std::complex<long double> t = z;
+  std::complex<long double> d = t;
 
   while (iterCount < MAX_ITERS_COUNT && abs(z) < maxT && abs(d) > minT) {
     z = z - sin(z) / cos(z);
@@ -45,10 +45,10 @@ uint32_t getColor( double zx,
 //////////////////////////////////////////////////////////////////////////
 
 void
-NewtonResetBounds(double *lx,
-                  double *rx,
-                  double *ty,
-                  double *by) {
+NewtonResetBounds(long double *lx,
+                  long double *rx,
+                  long double *ty,
+                  long double *by) {
   *lx = -4.15;
   *by = -1.0;
   *rx = *ty = 1.0;
@@ -56,9 +56,9 @@ NewtonResetBounds(double *lx,
 ///////////////////////////////////////////////////////
 
 void
-NewtonFillLine(double lx,
-               double dx,
-               double y,
+NewtonFillLine(long double lx,
+               long double dx,
+               long double y,
                uint32_t yix,
                uint32_t width,
                uint32_t *dst) {
